Remove all Passive Traits from Quick Actions
Passive Traits are those that require no action. Even if they have an activity there is no reason to display them, for instance "Aspect of the Wilds" will likely be configured with an activity, but it should AUTOMATICALLY run rather than be triggered by the Player / GM, therefore it should not appear.

Signed-off-by: aljames-arctic <bakanabaka.vtt@gmail.com>