nit: Reverse order of sort
If an item has multiple activation types, the one with the lowest sort value is prioritized (ie the one with the highest action cost).

This means a weapon that can be swung as an action or be used to channel magic with a bonus action will show up in the action slot.

Signed-off-by: aljames-arctic <bakanabaka.vtt@gmail.com>